+++ ocfl-java-api/src/main/java/io/ocfl/api/model/OcflObjectVersionFile.java
@@ -27,6 +27,7 @@ package io.ocfl.api.model;
 import io.ocfl.api.OcflFileRetriever;
 import io.ocfl.api.io.FixityCheckInputStream;
 import io.ocfl.api.util.Enforce;
+import java.io.InputStream;
 import java.util.Map;
 
 /**
@@ -79,6 +80,21 @@ public class OcflObjectVersionFile {
      */
     public FixityCheckInputStream getStream() {
         return fileRetriever.retrieveFile();
+    }
+
+    /**
+     * Returns an input stream of the file's content between the specified byte range. startPosition and endPosition
+     * may be null, depending on the underlying implementation, and the meaning of a null value is also implementation
+     * dependent.
+     *
+     * <p>The caller is responsible for closing the stream. The input stream is buffered.
+     *
+     * @param startPosition the byte offset in the file to start reading, inclusive
+     * @param endPosition the byte offset in the file to stop reading, inclusive
+     * @return a buffered input stream containing the specified file data
+     */
+    public InputStream getRange(Long startPosition, Long endPosition) {
+        return fileRetriever.retrieveRange(startPosition, endPosition);
     }
 
     @Override